+++ docs/Whats-New.md
@@ -294,6 +294,7 @@ New:
 - Single-color weapon lasers (by Starkku)
 - Customizable projectile trajectory (by secsome)
 - Correct owner house for Warhead Anim/SplashList & Play Animation trigger animations (by Starkku)
+- Customizable FLH When Infantry Is Crouched Or Deployed (by FS-21)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)